Extract the Audit module (audit logging) into Api/Audit
Third module of the #777 folder migration. Moves the append-only SSO audit
logger into Jellyfin.Plugin.SSO_Auth.Api.Audit:

- SsoAudit -> Api/Audit/ (namespace Api.Audit), import added at each call site.

SsoAudit is a leaf: it depends only on the logging abstraction and its own
event strings, no other Api module. Registers the Audit case on the
ApiModule_IsALeaf_ImportsNoOtherApiModule theory.

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1574 both TFMs), every
existing conformance rule, and the ABI-floor build.

Refs #777
