+++ SSO-Auth/Config/ProviderConfigStore.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Audit;
 using MediaBrowser.Model.Plugins;
 using Microsoft.Extensions.Logging;
